Pulling all alias-related annotation code strictly into paths.
