+++ tests/Parser/Annotations/ContentTypeAnnotationTest.php
@@ -37,8 +37,6 @@ class ContentTypeAnnotationTest extends AnnotationTest
         } else {
             $this->assertFalse($annotation->getVersion());
         }
-
-        $this->assertEmpty($annotation->getAliases());
     }
 
     public function providerAnnotation(): array
