+++ tests/Parser/Annotations/DescriptionAnnotationTest.php
@@ -29,7 +29,6 @@ class DescriptionAnnotationTest extends AnnotationTest
         $this->assertSame($expected, $annotation->toArray());
         $this->assertEmpty($annotation->getVendorTags());
         $this->assertFalse($annotation->getVersion());
-        $this->assertEmpty($annotation->getAliases());
     }
 
     public function providerAnnotation(): array
