+++ tests/Parser/Annotations/ErrorAnnotationTest.php
@@ -56,8 +56,6 @@ class ErrorAnnotationTest extends AnnotationTest
         } else {
             $this->assertFalse($annotation->getVersion());
         }
-
-        $this->assertEmpty($annotation->getAliases());
     }
 
     public function providerAnnotation(): array
