+++ tests/Parser/Annotations/MinVersionAnnotationTest.php
@@ -30,7 +30,6 @@ class MinVersionAnnotationTest extends AnnotationTest
         $this->assertSame($expected, $annotation->toArray());
         $this->assertEmpty($annotation->getVendorTags());
         $this->assertFalse($annotation->getVersion());
-        $this->assertEmpty($annotation->getAliases());
     }
 
     public function providerAnnotation(): array
