+++ tests/Parser/Annotations/ParamAnnotationTest.php
@@ -65,8 +65,6 @@ class ParamAnnotationTest extends AnnotationTest
         } else {
             $this->assertFalse($annotation->getVersion());
         }
-
-        $this->assertEmpty($annotation->getAliases());
     }
 
     public function providerAnnotation(): array
